services/ai/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.26.0

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Vector & Graph
qdrant-client>=1.7.0
asyncpg>=0.29.0

# Numerical
numpy>=1.26.0
tiktoken>=0.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
bandit>=1.7.0
hypothesis>=6.0.0

# Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp

# Workflow & Messaging
temporalio
nats-py>=2.6.0

# gRPC (Phase 4)
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Tree-sitter (Tier 0 verification)
tree-sitter>=0.21.0
tree-sitter-languages>=1.10.0

# WASM Sandbox (optional)
# wasmtime>=17.0.0

# Formal Verification (Tier 3)
# z3-solver>=4.12.0
